Fake stack outputs in cloudformation create check-mode
When run in check mode for a stack that would be created, we
previously omitted the outputs hash - the stack doesn't exist yet, so
we can't get its outputs.  This causes a problem when the
cloudformation task registers a variable which is then used by later
tasks in the playbook.

This commit looks at the outputs defined in the cloudformation
template and builds a corresponding outputs hash.  The value of each
output is a placeholder containing the stack and output names.  When
these outputs are used as parameters to other tasks, the placeholders
stand out in the check-mode output.